Modularize and clean the codebase.
